module 配置を整理
