+++ chapter10/src/10_2_asm_syscall/linux/arch/x86_64/fcntl.rs
@@ -1,23 +1,7 @@
-use std::path::Path;
+mod fd;
+mod flock;
+mod open;
 
-use crate::linux;
-
-/// ファイルディスクリプタ
-#[derive(Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
-pub(crate) struct Fd(i32);
-
-/// `open(2)` システムコールで、第2引数 `flags` を `O_RDONLY` にしたもの。
-pub(crate) fn open_readonly<P: AsRef<Path>>(path: P) -> linux::Result<Fd> {
-    todo!()
-}
-
-#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
-pub(crate) enum FlockOperation {
-    LockExclusive,
-    Unlock,
-}
-
-/// `flock(2)` に対応。
-pub(crate) fn flock(fd: &Fd, operation: FlockOperation) -> linux::Result<()> {
-    todo!()
-}
+pub(crate) use fd::Fd;
+pub(crate) use flock::{flock, FlockOperation};
+pub(crate) use open::open_readonly;
