+++ chapter10/src/10_2_asm_syscall/linux/arch/x86_64/fcntl/fd.rs
@@ -1,0 +1,5 @@
+/// ファイルディスクリプタ
+#[derive(Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
+pub(crate) struct Fd(i32);
+
+// TODO Drop で close
