+++ chapter10/src/10_2_asm_syscall/linux/arch/x86_64/fcntl/flock.rs
@@ -1,0 +1,14 @@
+use crate::linux;
+
+use super::fd::Fd;
+
+#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
+pub(crate) enum FlockOperation {
+    LockExclusive,
+    Unlock,
+}
+
+/// `flock(2)` に対応。
+pub(crate) fn flock(fd: &Fd, operation: FlockOperation) -> linux::Result<()> {
+    todo!()
+}
