+++ chapter10/src/10_2_asm_syscall/linux/arch/x86_64/fcntl/open.rs
@@ -1,0 +1,10 @@
+use std::path::Path;
+
+use crate::linux;
+
+use super::Fd;
+
+/// `open(2)` システムコールで、第2引数 `flags` を `O_RDONLY` にしたもの。
+pub(crate) fn open_readonly<P: AsRef<Path>>(path: P) -> linux::Result<Fd> {
+    todo!()
+}
